Patched scintilla
Copied/removed files as told by:
    contrib/patches/scintilla/__patchinfo__.py

Patched files in order, as told by:
    contrib/patches/scintilla/series

NOTE: The patch contrib/patches/scintilla/win32_build_stdvector.patch
fails to patch non-existent file, patch needs edition and changing:
src\CellBuffer.cxx -> src/CellBuffer.cxx

+++ contrib/scintilla/lexlib/WordList.cxx
@@ -191,7 +191,7 @@ bool WordList::InListAbbreviated(const char *s, const char marker) const {
 	if (0 == words)
 		return false;
 	unsigned char firstChar = s[0];
-	int j = starts[firstChar];
+	int j = starts[(int)firstChar];
 	if (j >= 0) {
 		while (static_cast<unsigned char>(words[j][0]) == firstChar) {
 			bool isSubword = false;
